Setup, HTML skeleton
In this commit, I have undergone the setup process
for this repository, including a file structure,
now with consideration for usage of webpack. I
have also set up an HTML skeleton to lay the
foundation of my project. In further commits, most
of the rest of the content will be set up using js
with CSS styling.