Complete setup w/ webpack
In this commit, I have finished setting up the
necessary tools in webpack.config.js and
package.json which will enable a smooth dev
process for this project. I also ran an initial
functionality test in the browser to ensure that
my "watch" and "build" scripts are working as
expected.