First draft homepage layout
In this commit, I've begun drafting the homepage
layout within src/index.html, with styles and
assets added in the stylesheet. This layout will
later be created entirely in JS, but I decided to
draft this in the HTML first so that I can get an
idea for what I want it to look like before
trapping myself into editing copious JS code to
make small changes to HTML content.

This portion of the initial layout includes:
- fonts and animated background assets
- all text content for homepage
- few applied styles via CSS