First draft nav styling
In this commit, I've added style properties to the
nav content for the homepage in my first draft
(created using HTML rather than JS), using a
combination of flexbox and grid, and colors picked
from the gif I'm using for the background of the
page. Currently having issues rendering the icon
images, but looking into troubleshooting that in
the next commit.