Fix syntax error re: socials icons
In this commit, I fixed minor syntactical errors
on the lines sourcing the images for the socials
icons in the nav bar. I also applied the icon
class to each of these images, as well as their
link elements, in order to properly size the icons
within the nav bar.